deleteForm
delete form middleware, handler,controller and test

+++ src/controllers/formControllers/deleteFormController.js
@@ -1,11 +1,16 @@
 const prisma = require("../../db.js");
 
 const deleteFormController = async (data) => {
-  const form = await prisma.form.delete({
-    where: {
-      form_id: data.form_id,
-    },
+  const form = await prisma.form.findUnique({
+    where: { form_id: data.form_id },
   });
+if(!form) return {status:404, message: "Form not found."};
+
+await prisma.form.delete({
+  where: { form_id: data.form_id },
+});
+
+
   return { status: 200, message: "Form deleted successfully", form: form };
 };
